requirements.txt
fastapi>=0.100
uvicorn>=0.23
python-multipart>=0.0.7
pydantic>=2.0
beautifulsoup4>=4.12
joblib>=1.3
scikit-learn>=1.3
extract-msg>=0.45
python-whois>=0.8.0
